+++ CHANGELOG.md
@@ -8,6 +8,9 @@ documented in this file.
 ## 1.2.0
 
 - Dropped PHP 8.0 support, added testing for 8.4. New minimum PHP version is 8.1.
+- Renamed `is_missing_class()` to `isMissingClass()` and `get_apcu_prefix()` to `getApcuPrefix()`.
+- Deprecated `set_apcu_prefix()` but method still exists for backwards
+  compatibility. `setApcuPrefix()` is the new method name.
 
 ## 1.1.1 - 2022-08-31